refactor(artifact-provenance): extract message finalization
